Fixing some odd issue of character sets!
I don’t think I’ve seen this before!  Somehow, there is a difference in
the character sets.

I had no problem locally on my Mac, but it failed on the website.

I solved by copying the page name from the website and pasting back
into the demo doc.  Looks exactly the same as the no-good link, but in
git you can see that the é is understood as different.