Hmm, just noticed that I can collect comments in the commit window on the Github client AS I’M DOING THE WORK.
Having comments here may not be the best place, but might be better
then in the files themselves, as I’ve done in past.

So:  notice that the Definitions are i) in the back and ii) mostly
cross-references to places in the document where the terms are defined.
 A few thoughts on that: — Definitions in the back seems like a good
idea to me even though it is common to put them in the front.
Definitions are - or should become sort of background or environmental
- part of what is assumed.  They should be relatively stable. On the
other hand, this is in tension with another use of Definitions - which
is as a kind of “parameter”.  E.g. “Seller will pay the Price.”  and
“Price means $102.”   So, to be considered and worked out.  ii)
embedding definitions in text is convenient for the drafter, but often
less convenient for the reader, and less scalable.