NDA - FR - Bilatéral - From Me. Céline Bonnard
Rough.  Note that in using Z/ol/s15, we end up with ugly hanging
numbers.  There are three ways to fix this.  i)  change the numbers of
the subparagraphs so that they start with 0 instead of 1 - 1.0.sec
instead of 1.1.sec and renumber.  ii) add titles and use Z/ol/15.  iii)
make a new widget that uses <p> or <br> instead of <li>.  To be
considered.