CabinetBondard - New Docs put into Folders - and renaming the NDAs
Attempt to put all into a general tree using English names for the
kinds of documents.  This is only a suggestion, and it may be that we
will decide to do this by French names.  The idea is that this makes it
easier to create uniformity across various jurisdictions and languages.
 Rather like how HTML and software work.  Part of the pragmatic
imperialism of English.